Extract BundleManager to an Interface in SamzaRunner (#26268)
