+++ runners/samza/src/main/java/org/apache/beam/runners/samza/runtime/DoFnOp.java
@@ -200,7 +200,7 @@ public class DoFnOp<InT, FnOutT, OutT> implements Op<InT, OutT, Void> {
     final FutureCollector<OutT> outputFutureCollector = createFutureCollector();
 
     this.bundleManager =
-        new BundleManager<>(
+        new ClassicBundleManager<>(
             createBundleProgressListener(),
             outputFutureCollector,
             samzaPipelineOptions.getMaxBundleSize(),
